Add spring data mongodb reactive dependency - Remove spring data r2dcb, mysql-connector-j, r2dbc-mysql and  spring boot devtools dependency - Add docker compose file for keycloak and mongodb - Modify domain model NourishmentType with sealed and records

+++ nourishment/src/main/java/me/kirenai/re/nourishment/domain/model/dto/ListNourishmentsTypePercentageResponse.java
@@ -5,19 +5,19 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.Builder;
-import me.kirenai.re.nourishment.util.enums.NourishmentTypeEnum;
+import me.kirenai.re.nourishment.domain.enums.NourishmentTypeEnum;
 
 @Builder
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonDeserialize(as = ListNourishmentsTypePercentageResponse.class)
 public record ListNourishmentsTypePercentageResponse(
         NourishmentTypeEnum nourishmentType,
-        Integer percentage
+        Double percentage
 ) implements ListNourishmentsTypeResponse {
 
     @JsonCreator
     public ListNourishmentsTypePercentageResponse(@JsonProperty("nourishmentType") NourishmentTypeEnum nourishmentType,
-                                                  @JsonProperty("percentage") Integer percentage) {
+                                                  @JsonProperty("percentage") Double percentage) {
         this.nourishmentType = NourishmentTypeEnum.PERCENTAGE;
         this.percentage = percentage;
     }
